+++ apps/app/src/test/scala/com/daml/network/integration/tests/SvOffboardingIntegrationTest.scala
@@ -37,6 +37,7 @@ class SvOffboardingIntegrationTest extends SvIntegrationTestBase {
           action,
           "url",
           "description",
+          sv1Backend.getSvcInfo().svcRules.payload.config.voteRequestTimeout,
         )
       },
     )(