+++ apps/app/src/test/scala/com/daml/network/util/ConfigScheduleUtil.scala
@@ -125,6 +125,7 @@ trait ConfigScheduleUtil extends CNNodeTestCommon {
             action,
             "url",
             "description",
+            sv1Backend.getSvcInfo().svcRules.payload.config.voteRequestTimeout,
           )
         },
       )(